+++ CHANGELOG.rst
@@ -185,6 +185,12 @@ Changelog
 0.4.6 (2017-05-XX)
 ------------------
 
-* Added a `number_of_lines` parameter for the file I\O process. Keep at `-1` to read everything.
+* Added a `number_of_lines` parameter for the file I\O process. Keep at `-1` to read everything. Controlled via the "Debug?" field in the GUI.
 
 * Changed defaults in GUI.
+
+* Allows acquisition in bi-directional scanning mode.
+
+* Back-end changes for possible future support of binary files.
+
+* The code allows to dismiss unwanted input channels by specifying them as "Empty". If you fail to